Better timeout tests
