+++ src/test/java/com/stripe/functional/RequestOptionsTest.java
@@ -4,19 +4,13 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 import com.stripe.BaseStripeTest;
-import com.stripe.Stripe;
-import com.stripe.exception.APIConnectionException;
 import com.stripe.exception.StripeException;
 import com.stripe.model.Balance;
 import com.stripe.net.RequestOptions;
 import com.stripe.net.StripeResponse;
 
-import java.io.IOException;
-
 import org.junit.Ignore;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 
 public class RequestOptionsTest extends BaseStripeTest {
   @Ignore // stripe-mock doesn't send a Stripe-Version header
@@ -45,34 +39,5 @@ public class RequestOptionsTest extends BaseStripeTest {
 
     assertNotNull(response);
     assertEquals(idempotencyKey, response.headers().get("Idempotency-Key"));
-  }
-
-  @Rule
-  public ExpectedException thrown = ExpectedException.none();
-
-  @Test
-  public void testConnectTimeout() throws IOException, StripeException {
-    // Kind of terrible, but let's test connection timeouts with an external server and a super
-    // short timeout.
-    Stripe.overrideApiBase("https://api.stripe.com");
-
-    thrown.expect(APIConnectionException.class);
-    thrown.expectMessage("connect timed out");
-
-    final RequestOptions options = RequestOptions.builder().setConnectTimeout(1).build();
-    Balance.retrieve(options);
-  }
-
-  @Test
-  public void testReadTimeout() throws IOException, StripeException {
-    // Also kind of terrible, but let's test read connection timeouts with a valid external server
-    // but a super short timeout.
-    Stripe.overrideApiBase("https://api.stripe.com");
-
-    thrown.expect(APIConnectionException.class);
-    thrown.expectMessage("Read timed out");
-
-    final RequestOptions options = RequestOptions.builder().setReadTimeout(1).build();
-    Balance.retrieve(options);
   }
 }
